BCM-1365: Modify channel events API to include querying context info (#65)
* BCM-1365: Modify channel events API to include querying context info

* BCM-1365: update crec-api-go version

+++ services/watchers/watchers.go
@@ -107,7 +107,7 @@ type UpdateWatcherInput struct {
 
 type WatcherFilters struct {
 	Limit         *int           `url:"limit,omitempty"`
-	Offset        *int           `url:"offset,omitempty"`
+	Offset        *int64         `url:"offset,omitempty"`
 	Name          *string        `url:"name,omitempty"`
 	Status        *WatcherStatus `url:"status,omitempty"`
 	ChainSelector *string        `url:"chain_selector,omitempty"`
